Cleanup issue for branch with missing merge point
This fixes an issue where you have a tree of branches like so

main
  test
    test2

and you git branch -D test. The git config still has the merge point
set to test but it doesn't exist. This throws an error when you run
rev-list on a missing branch. Guard against that here.