Fix MaxDepth
This fixes a bug with MaxDepth that was also causing a panic because
we expected the depth to be one more than it was returning. This fixes
the function to now return the depth that is expected and in the case
any more bugs exist in drawing the graph we handle that with an explicit
check to ensure the padding is never less than 0.